Fix up Url Checker defaulting.
