+++ tests/TestCase/Authenticator/EnvironmentAuthenticatorTest.php
@@ -378,7 +378,7 @@ class EnvironmentAuthenticatorTest extends TestCase
         $envAuth = new EnvironmentAuthenticator($this->identifier, [
             'loginUrl' => '%^/[a-z]{2}/users/secure/?$%',
             'urlChecker' => [
-                'className' => 'Authentication.Default',
+                'className' => 'Authentication.Generic',
                 'useRegex' => true,
             ],
             'fields' => [
@@ -409,7 +409,7 @@ class EnvironmentAuthenticatorTest extends TestCase
         $envAuth = new EnvironmentAuthenticator($this->identifier, [
             'loginUrl' => '%auth\.localhost/[a-z]{2}/users/secure/?$%',
             'urlChecker' => [
-                'className' => 'Authentication.Default',
+                'className' => 'Authentication.Generic',
                 'useRegex' => true,
                 'checkFullUrl' => true,
             ],
@@ -441,7 +441,7 @@ class EnvironmentAuthenticatorTest extends TestCase
         $envAuth = new EnvironmentAuthenticator($this->identifier, [
             'loginUrl' => '%auth\.localhost/[a-z]{2}/users/secure/?$%',
             'urlChecker' => [
-                'className' => 'Authentication.Default',
+                'className' => 'Authentication.Generic',
                 'useRegex' => true,
                 'checkFullUrl' => true,
             ],
